Move attribute functions into AttributeRepository instead of CombinationRepository
