+++ src/Adapter/Attribute/Repository/AttributeRepository.php
@@ -30,6 +30,7 @@ namespace PrestaShop\PrestaShop\Adapter\Attribute\Repository;
 
 use Doctrine\DBAL\Connection;
 use PrestaShop\PrestaShop\Adapter\AbstractObjectModelRepository;
+use PrestaShop\PrestaShop\Core\Domain\Language\ValueObject\LanguageId;
 use PrestaShop\PrestaShop\Core\Domain\Product\AttributeGroup\Attribute\Exception\AttributeNotFoundException;
 use RuntimeException;
 
@@ -81,5 +82,93 @@ class AttributeRepository extends AbstractObjectModelRepository
         if (count($attributeIds) !== $result) {
             throw new AttributeNotFoundException('Some of provided attributes does not exist');
         }
+    }
+
+    /**
+     * @param int[] $combinationIds
+     * @param LanguageId $langId
+     *
+     * @return array<int, array<int, mixed>>
+     */
+    public function getAttributesInfoByCombinationIds(array $combinationIds, LanguageId $langId): array
+    {
+        $attributeCombinationAssociations = $this->getAttributeCombinationAssociations($combinationIds);
+
+        $attributeIds = array_unique(array_map(function (array $attributeByCombination): int {
+            return (int) $attributeByCombination['id_attribute'];
+        }, $attributeCombinationAssociations));
+
+        $attributesInfoByAttributeId = $this->getAttributesInformation($attributeIds, $langId->getValue());
+
+        $attributesInfoByCombinationId = [];
+        foreach ($attributeCombinationAssociations as $attributeCombinationAssociation) {
+            $combinationId = (int) $attributeCombinationAssociation['id_product_attribute'];
+            $attributeId = (int) $attributeCombinationAssociation['id_attribute'];
+            $attributesInfoByCombinationId[$combinationId][] = $attributesInfoByAttributeId[$attributeId];
+        }
+
+        return $attributesInfoByCombinationId;
+    }
+
+    /**
+     * @param int[] $combinationIds
+     *
+     * @return array<int, array<string, mixed>>
+     */
+    private function getAttributeCombinationAssociations(array $combinationIds): array
+    {
+        $qb = $this->connection->createQueryBuilder();
+        $qb->select('pac.id_attribute')
+            ->addSelect('pac.id_product_attribute')
+            ->from($this->dbPrefix . 'product_attribute_combination', 'pac')
+            ->where($qb->expr()->in('pac.id_product_attribute', ':combinationIds'))
+            ->setParameter('combinationIds', $combinationIds, Connection::PARAM_INT_ARRAY)
+        ;
+
+        return $qb->execute()->fetchAll();
+    }
+
+    /**
+     * @param int[] $attributeIds
+     * @param int $langId
+     *
+     * @return array<int, array<int, mixed>>
+     */
+    private function getAttributesInformation(array $attributeIds, int $langId): array
+    {
+        $qb = $this->connection->createQueryBuilder();
+        $qb->select('a.id_attribute')
+            ->addSelect('ag.id_attribute_group')
+            ->addSelect('al.name AS attribute_name')
+            ->addSelect('agl.name AS attribute_group_name')
+            ->from($this->dbPrefix . 'attribute', 'a')
+            ->leftJoin(
+                'a',
+                $this->dbPrefix . 'attribute_lang',
+                'al',
+                'a.id_attribute = al.id_attribute AND al.id_lang = :langId'
+            )->leftJoin(
+                'a',
+                $this->dbPrefix . 'attribute_group',
+                'ag',
+                'a.id_attribute_group = ag.id_attribute_group'
+            )->leftJoin(
+                'ag',
+                $this->dbPrefix . 'attribute_group_lang',
+                'agl',
+                'agl.id_attribute_group = ag.id_attribute_group AND agl.id_lang = :langId'
+            )->where($qb->expr()->in('a.id_attribute', ':attributeIds'))
+            ->setParameter('attributeIds', $attributeIds, Connection::PARAM_INT_ARRAY)
+            ->setParameter('langId', $langId)
+        ;
+
+        $attributesInfo = $qb->execute()->fetchAll();
+
+        $attributesInfoByAttributeId = [];
+        foreach ($attributesInfo as $attributeInfo) {
+            $attributesInfoByAttributeId[(int) $attributeInfo['id_attribute']] = $attributeInfo;
+        }
+
+        return $attributesInfoByAttributeId;
     }
 }
