+++ src/PrestaShopBundle/Resources/config/services/adapter/product/_combination.yml
@@ -50,6 +50,7 @@ services:
             - '@prestashop.core.util.number.number_extractor'
             - '@prestashop.adapter.product.stock.repository.stock_available_repository'
             - '@prestashop.core.grid.query_builder.product_combination'
+            - '@prestashop.adapter.attribute.repository.attribute_repository'
         tags:
             - name: tactician.handler
               command: PrestaShop\PrestaShop\Core\Domain\Product\Combination\Query\GetEditableCombinationsList
